requirements-tests.txt
asynctest
pyyaml
